Saving

+++ src/main/java/org/legendofdragoon/fonteditor/MathHelper.java
@@ -17,7 +17,7 @@ public final class MathHelper {
     return value;
   }
 
-  public static void set(final byte[] data, final int offset, final int size, final long value) {
+  public static void set(final byte[] data, final int offset, final int size, final int value) {
     for(int i = 0; i < size; i++) {
       data[offset + i] = (byte)(value >>> i * 8 & 0xff);
     }
